Implement wake up brightness and transition

+++ src/hue/api/behavior.rs
@@ -69,7 +69,7 @@ pub mod configuration {
 
     #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
     pub struct FadeInDuration {
-        pub seconds: i64,
+        pub seconds: u32,
     }
 
     #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
